refactor: 훅 사용 컴포넌트 위치 변경
