+++ src/features/auth/ui/UserInfoPresentation.tsx
@@ -1,25 +1,20 @@
 import React from "react";
-import jwt from "jsonwebtoken";
 
 import { Button, Stack, Text } from "@chakra-ui/react";
 import { ProgressCircleRing, ProgressCircleRoot } from "@/shared/chakra-ui/progress-circle";
 
-export interface Parameters extends jwt.JwtPayload {
-    isLoggedIn: boolean;
-    onLogoutClick: () => void;
-}
+import useUserInfo from "../lib/useUserInfo";
+import getLoggedIn from "../lib/getLoggedIn";
 
-export default function UserInfoPResentation({
-    "cognito:username": username,
-    email,
-    isSuccess,
-    onLogoutClick,
-}: Parameters): React.ReactElement {
+export default function UserInfoPResentation(): React.ReactElement {
+    const { "cognito:username": username, email } = useUserInfo();
+    const { isLoggedIn, onLogout } = getLoggedIn();
+
     return (
         <>
-            {isSuccess ? (
+            {isLoggedIn ? (
                 <Stack w="100%">
-                    <Button onClick={onLogoutClick}>로그아웃</Button>
+                    <Button onClick={onLogout}>로그아웃</Button>
                     <Text>아이디: {username}</Text>
                     <Text>이메일: {email}</Text>
                 </Stack>
